Disable CSR utilities on K8s <1.19
Fixes #816
